+++ ptn/missionalertbot/modules/TextGen.py
@@ -23,7 +23,7 @@ def txt_create_discord(mission_params):
         f"{mission_params.commodity_name} "
         f"{'from' if mission_params.mission_type == 'load' else 'to'} **{mission_params.station.upper()}** station in system "
         f"**{mission_params.system.upper()}** : {mission_params.profit}k per unit profit : "
-        f"{mission_params.demand} {'demand' if mission_params.mission_type == 'load' else 'supply'} : {mission_params.pads.upper()}-pads."
+        f"{mission_params.demand}k {'demand' if mission_params.mission_type == 'load' else 'supply'} : {mission_params.pads.upper()}-pads."
     )
     return discord_text
 
@@ -44,7 +44,7 @@ def txt_create_reddit_body(mission_params):
             f"\n\n"
             f"**BUY FROM**: station **{mission_params.station.upper()}** ({mission_params.pads.upper()}-pads) in system **{mission_params.system.upper()}**\n\n**COMMODITY**: "
             f"{mission_params.commodity_name}\n\n&#x200B;\n\n**SELL TO**: Fleet Carrier **{mission_params.carrier_data.carrier_long_name} "
-            f"{mission_params.carrier_data.carrier_identifier}**\n\n**PROFIT**: {mission_params.profit}k/unit : {mission_params.demand} "
+            f"{mission_params.carrier_data.carrier_identifier}**\n\n**PROFIT**: {mission_params.profit}k/unit : {mission_params.demand}k "
             f"demand\n\n\n\n[Join us on Discord]({constants.REDDIT_DISCORD_LINK_URL}) for "
             f"mission updates and discussion, channel **#{mission_params.carrier_data.discord_channel}**.")
     else:
@@ -54,6 +54,6 @@ def txt_create_reddit_body(mission_params):
             f"**BUY FROM**: Fleet Carrier **{mission_params.carrier_data.carrier_long_name} {mission_params.carrier_data.carrier_identifier}**"
             f"\n\n**COMMODITY**: {mission_params.commodity_name}\n\n&#x200B;\n\n**SELL TO**: station "
             f"**{mission_params.station.upper()}** ({mission_params.pads.upper()}-pads) in system **{mission_params.system.upper()}**\n\n**PROFIT**: {mission_params.profit}k/unit "
-            f": {mission_params.demand} supply\n\n\n\n[Join us on Discord]({constants.REDDIT_DISCORD_LINK_URL}) for mission updates"
+            f": {mission_params.demand}k supply\n\n\n\n[Join us on Discord]({constants.REDDIT_DISCORD_LINK_URL}) for mission updates"
             f" and discussion, channel **#{mission_params.carrier_data.discord_channel}**.")
     return reddit_body